\name{DangLoop-package}
\alias{DangLoop-package}
\alias{DangLoop}
\docType{package}
\title{\packageTitle{DangLoop}}
\description{\packageDescription{DangLoop}}
\section{Package Content}{\packageIndices{DangLoop}}
\author{\packageAuthor{DangLoop}}
\section{Maintainer}{\packageMaintainer{DangLoop}}
\keyword{package}
